后台管理

+++ src/pages/layout.vue
@@ -4,34 +4,19 @@
     <ts-navbar :userInfo="userInfo" :userMenu="userMenu"  :default-active="activeIndex" @sidebarHide="hideside" @userItemClick="useritemclick">
       <router-view></router-view>
     </ts-navbar>
-    <el-dialog title="修改密码" :visible.sync="modifyDialog" width="380px" @close="resetForm('modify')">
-      <el-form :model="modify" label-position="right" label-width="80px" :rules="rules" ref="modify" class="marginXauto">
-          <el-form-item label="原密码" prop="oldPassword">
-            <el-input v-model="modify.oldPassword" class="width250"></el-input>
-          </el-form-item>
-          <el-form-item label="新密码" prop="newPassword">
-            <el-input v-model="modify.newPassword" class="width250"></el-input>
-          </el-form-item>
-          <el-form-item label="确认密码" prop="newPassword2">
-            <el-input v-model="modify.newPassword2" class="width250"></el-input>
-          </el-form-item>
-          <div style="padding-left:100px;">
-          <el-button type="primary" @click="modifyPass">确定</el-button>
-          <el-button @click="resetForm('modify')">取消</el-button>
-          </div>
-      </el-form>
-    </el-dialog>
   </div>
 </template>
 <script>
 import { sideRoutes } from '@/router/routes.js'
 import userphoto from '@/assets/logo.png'
+import axion from '@/util/api.js'
 // import logo from '@/assets/img/default/hosLogo.png'
 // import simplelogo from '@/assets/img/default/simplelogo.png'
 // const ifnew = window.location.href.indexOf('new');
 export default {
   data() {
     return {
+      token:sessionStorage.getItem('token'),
       isCollapse: false,
       routeMenu: sideRoutes,
       userInfo: {
@@ -39,8 +24,6 @@ export default {
         username: '个人中心'
       },
       userMenu: [{
-        text: '修改密码'
-      }, {
         text: '退出'
       }],
       activeIndex: '1',
@@ -68,51 +51,40 @@ export default {
       }
     }
   },
+  mounted() {
+    this.getUserInfo()
+  },
   methods: {
     hideside(isCollapse) {
       this.isCollapse = isCollapse
     },
     useritemclick(i) {
       if (i.text == '退出') {
-        this.$message(`您选择了${i.text}`)
-        sessionStorage.clear();
-        localStorage.clear();
-        this.$router.push('/')
-      }
-      if (i.text == '修改密码') {
-        this.modifyDialog = true
+        let param = {
+          token:this.token
+        }
+        axion.logout(param).then(res => {
+          if(res.data.retCode == 0) {
+            this.$message(`您选择了${i.text}`)
+            sessionStorage.clear();
+            this.$router.push('/')
+          }
+        })
       }
     },
-    //重置表单
-    resetForm(formName){
-      this.$refs[formName].resetFields();
-      this.modifyDialog = false
-    },
-    //修改密码
-    modifyPass(){
-      this.setPasswordPrams.password = this.modifyPass.oldPassword
-      this.setPasswordPrams.newPassword = this.modify.newPassword
-      if(this.modify.oldPassword !="" && this.modify.newPassword != "" && this.modify.newPassword2 != "") {
-        if(this.modify.oldPassword == this.modify.newPassword) {
-          this.$message({
-						message: '新密码与当前密码不能相同',
-						type: 'warning'
-					});
-        }else if(this.modify.newPassword != this.modify.newPassword2){
-          this.$message({
-						message: '两次新密码输入不一样',
-						type: 'warning'
-					});
+    getUserInfo(){
+        let param = {
+            token:this.token
         }
-        else {
-          //调修改密码接口
-        }
-      }else {
-        this.$message({
-					message: '原密码、新密码和确认密码不能为空',
-					type: 'warning'
-				});
-      }
+        axion.getUserInfo(param).then( res => {
+            if(res.data.retCode == 0) {
+                this.userInfo.userphoto = 'http://192.168.1.101:8515/img/getPicture?path='+res.data.param.profilePic
+                console.log('11',this.userInfo.profilePic)
+            }else if(res.data.retCode == 50004) {
+                sessionStorage.removeItem('user_token');
+                this.$router.push({ path:'/'})
+            }
+        })
     },
   }
 }
